internal_link test [nfc]: Re-home narrow link tests
