+++ test/model/internal_link_test.dart
@@ -35,6 +35,100 @@ Future<PerAccountStore> setupStore({
 }
 
 void main() {
+  group('narrowLink', () {
+    test('CombinedFeedNarrow', () {
+      final store = eg.store();
+      check(narrowLink(store, const CombinedFeedNarrow()))
+        .equals(store.realmUrl.resolve('#narrow'));
+      check(narrowLink(store, const CombinedFeedNarrow(), nearMessageId: 1))
+        .equals(store.realmUrl.resolve('#narrow/near/1'));
+    });
+
+    test('MentionsNarrow', () {
+      final store = eg.store();
+      check(narrowLink(store, const MentionsNarrow()))
+        .equals(store.realmUrl.resolve('#narrow/is/mentioned'));
+      check(narrowLink(store, const MentionsNarrow(), nearMessageId: 1))
+        .equals(store.realmUrl.resolve('#narrow/is/mentioned/near/1'));
+    });
+
+    test('StarredMessagesNarrow', () {
+      final store = eg.store();
+      check(narrowLink(store, const StarredMessagesNarrow()))
+        .equals(store.realmUrl.resolve('#narrow/is/starred'));
+      check(narrowLink(store, const StarredMessagesNarrow(), nearMessageId: 1))
+        .equals(store.realmUrl.resolve('#narrow/is/starred/near/1'));
+    });
+
+    test('ChannelNarrow / TopicNarrow', () {
+      void checkNarrow(String expectedFragment, {
+        required int streamId,
+        required String name,
+        String? topic,
+        int? nearMessageId,
+      }) async {
+        assert(expectedFragment.startsWith('#'), 'wrong-looking expectedFragment');
+        final store = eg.store();
+        await store.addStream(eg.stream(streamId: streamId, name: name));
+        final narrow = topic == null
+          ? ChannelNarrow(streamId)
+          : TopicNarrow(streamId, topic);
+        check(narrowLink(store, narrow, nearMessageId: nearMessageId))
+          .equals(store.realmUrl.resolve(expectedFragment));
+      }
+
+      checkNarrow(streamId: 1,   name: 'announce',       '#narrow/stream/1-announce');
+      checkNarrow(streamId: 378, name: 'api design',     '#narrow/stream/378-api-design');
+      checkNarrow(streamId: 391, name: 'Outreachy',      '#narrow/stream/391-Outreachy');
+      checkNarrow(streamId: 415, name: 'chat.zulip.org', '#narrow/stream/415-chat.2Ezulip.2Eorg');
+      checkNarrow(streamId: 419, name: 'français',       '#narrow/stream/419-fran.C3.A7ais');
+      checkNarrow(streamId: 403, name: 'Hshs[™~}(.',     '#narrow/stream/403-Hshs.5B.E2.84.A2~.7D.28.2E');
+      checkNarrow(streamId: 60,  name: 'twitter', nearMessageId: 1570686, '#narrow/stream/60-twitter/near/1570686');
+
+      checkNarrow(streamId: 48,  name: 'mobile', topic: 'Welcome screen UI',
+                  '#narrow/stream/48-mobile/topic/Welcome.20screen.20UI');
+      checkNarrow(streamId: 243, name: 'mobile-team', topic: 'Podfile.lock clash #F92',
+                  '#narrow/stream/243-mobile-team/topic/Podfile.2Elock.20clash.20.23F92');
+      checkNarrow(streamId: 377, name: 'translation/zh_tw', topic: '翻譯 "stream"',
+                  '#narrow/stream/377-translation.2Fzh_tw/topic/.E7.BF.BB.E8.AD.AF.20.22stream.22');
+      checkNarrow(streamId: 42,  name: 'Outreachy 2016-2017', topic: '2017-18 Stream?', nearMessageId: 302690,
+                  '#narrow/stream/42-Outreachy-2016-2017/topic/2017-18.20Stream.3F/near/302690');
+    });
+
+    test('DmNarrow', () {
+      void checkNarrow(String expectedFragment, String legacyExpectedFragment, {
+        required List<int> allRecipientIds,
+        required int selfUserId,
+        int? nearMessageId,
+      }) {
+        assert(expectedFragment.startsWith('#'), 'wrong-looking expectedFragment');
+        final store = eg.store();
+        final narrow = DmNarrow(allRecipientIds: allRecipientIds, selfUserId: selfUserId);
+        check(narrowLink(store, narrow, nearMessageId: nearMessageId))
+          .equals(store.realmUrl.resolve(expectedFragment));
+        store.connection.zulipFeatureLevel = 176;
+        check(narrowLink(store, narrow, nearMessageId: nearMessageId))
+          .equals(store.realmUrl.resolve(legacyExpectedFragment));
+      }
+
+      checkNarrow(allRecipientIds: [1], selfUserId: 1,
+        '#narrow/dm/1-dm',
+        '#narrow/pm-with/1-pm');
+      checkNarrow(allRecipientIds: [1, 2], selfUserId: 1,
+        '#narrow/dm/1,2-dm',
+        '#narrow/pm-with/1,2-pm');
+      checkNarrow(allRecipientIds: [1, 2, 3], selfUserId: 1,
+        '#narrow/dm/1,2,3-group',
+        '#narrow/pm-with/1,2,3-group');
+      checkNarrow(allRecipientIds: [1, 2, 3, 4], selfUserId: 4,
+        '#narrow/dm/1,2,3,4-group',
+        '#narrow/pm-with/1,2,3,4-group');
+      checkNarrow(allRecipientIds: [1, 2], selfUserId: 1, nearMessageId: 12345,
+        '#narrow/dm/1,2-dm/near/12345',
+        '#narrow/pm-with/1,2-pm/near/12345');
+    });
+  });
+
   final realmUrl = Uri.parse('https://example.com/');
 
   void testExpectedNarrows(List<(String, Narrow?)> testCases, {
